+++ azure_devops_rust_api/src/git/models.rs
@@ -411,6 +411,8 @@ pub mod change {
         All,
         #[serde(rename = "delete, sourceRename")]
         DeleteSourceRename,
+        #[serde(rename = "edit, rename")]
+        EditRename,
     }
 }
 #[doc = ""]