+++ vsts-api-patcher/src/main.rs
@@ -605,8 +605,9 @@ impl Patcher {
             ["definitions", "Change", "properties", "changeType", "enum"] => {
                 println!("Update git Change changeType definition");
                 let mut value = value.clone();
-                // Add extra value discovered in testing.
+                // Add extra values discovered in testing.
                 value.push(JsonValue::from("delete, sourceRename")).unwrap();
+                value.push(JsonValue::from("edit, rename")).unwrap();
                 Some(value)
             }
             _ => None,